requirements.txt
# Core dependencies
Flask==2.3.2
Werkzeug==2.3.6
Jinja2==3.1.2
python-dotenv==1.0.0
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-Caching==2.0.2
Markdown==3.4.3

# Data processing
yfinance==0.2.18
pandas==2.0.3
numpy==1.25.0
plotly==5.15.0
matplotlib==3.7.2

# Financial analysis
alpaca-trade-api==3.0.0
textblob==0.17.1
newspaper3k==0.2.8

# Google Gemini API for AI
google-generativeai==0.3.1

# Production server
gunicorn==21.2.0
gevent==23.9.0

# Production monitoring and performance
Flask-Limiter==3.3.1
sentry-sdk==1.32.0

# Testing
pytest==7.3.1
pytest-cov==4.1.0

# Optional - Redis cache support
# redis==4.5.5